requirements.txt
python-dotenv==0.19.2
telegram==0.0.1
python-telegram-bot==13.11.0